# Improve speed of std.digest.digest!(Hash, Range) on non-array ranges