@kinke 2020-05-31
> While benchmarking I was surprised to discover that with the implementation prior to this PR dmd -O -inline was noticeably faster than ldc2 -O3 for CRC32 and every MurmurHash3 variant.

That should never happen. ;) - Please let us know via an LDC issue if you come across such a case. There's one catch though - due to cross-module inlining being disabled (and otherwise experimental), the code in question should first be tested with LTO (`-flto=<thin or full> -defaultlib=phobos2-ldc-lto,druntime-ldc-lto`), as that can indeed be a common bottleneck.